+++ World/Core/SplitConnectionTesselator.cpp
@@ -53,7 +53,7 @@ void SplitConnectionTesselator::JsonLoad(const rapidjson::Value & val)
 			{
 				PMesh<VertexVTN> m = std::make_shared<TemplateMesh<VertexVTN>>(*me);
 				const auto &uv4 = mTextures[6];
-				for (int i = 0; i < m->SizeVertex(); i++)
+				for (size_t i = 0; i < m->SizeVertex(); i++)
 				{
 					m->Vertex(i).texture.x = glm::mix(uv4.x, uv4.z, m->Vertex(i).texture.x);
 					m->Vertex(i).texture.y = glm::mix(uv4.y, uv4.w, m->Vertex(i).texture.y);
